+++ src/path_condition.mli
@@ -8,12 +8,17 @@ val pp : t Fmt.t
 
 val empty : t
 
-val add : Smtml.Typed.Bool.t -> t -> t
+(* You should only use this after checking the condition is SAT in the context of the path condition. If you add it without checking, it may lead to wrong results. *)
+val add_checked_sat_condition : Smtml.Typed.Bool.t -> t -> t
 
-(* CAUTION: this must only be called after the symbol has been added to the path condition *)
+(* Get all the slices that are related to the condition given in argument. The condition is not added to the result. *)
+val slice_on_new_condition : Smtml.Typed.Bool.t -> t -> Smtml.Expr.Set.t
+
+(* Get the slice for a symbol. Return an empty set if the symbol is not part of the path condition. *)
 val slice_on_symbol : Smtml.Symbol.t -> t -> Smtml.Expr.Set.t
 
-(* CAUTION: this must only be called after the condition added to the path condition with `add` *)
-val slice_on_condition : Smtml.Typed.Bool.t -> t -> Smtml.Expr.Set.t
+(* Get all slices of the path condition as a list. *)
+val to_list : t -> Smtml.Expr.Set.t list
 
-val slice : t -> Smtml.Expr.Set.t list
+(* Return a map of known equalities from symbols to values. *)
+val get_known_equalities : t -> Smtml.Value.t Smtml.Symbol.Map.t