+++ test/test_path_condition.ml
@@ -15,43 +15,59 @@ let print pc = Fmt.str "%a" Symex.Path_condition.pp pc
 let test_print_pc_empty () : unit =
   let pc = Symex.Path_condition.empty |> print in
   Alcotest.(check string)
-    "same string" "((aliases_of_canonicals {}) (payload_of_canonicals {}))" pc
+    "same string"
+    "union find:\n\
+    \  ((aliases_of_canonicals {}) (payload_of_canonicals {}))\n\
+     equalities:\n\
+    \  {}\n"
+    pc
 
 let test_print_pc_one () : unit =
   let pc =
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (i32_sym "x0") (i32_const 42l))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "x0") (i32_const 42l))
     |> print
   in
   Alcotest.(check string)
     "same string"
-    "((aliases_of_canonicals {(x0 {})})\n\
-    \ (payload_of_canonicals {(x0 (bool.eq x0 42))}))"
+    "union find:\n\
+    \  ((aliases_of_canonicals {}) (payload_of_canonicals {}))\n\
+     equalities:\n\
+    \  {x0 = 42}\n"
     pc
 
 let test_print_pc_two () : unit =
   let pc =
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (i32_sym "x0") (i32_const 42l))
-    |> Symex.Path_condition.add (eq (i32_sym "y0") (i32_const 56l))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "x0") (i32_const 42l))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "y0") (i32_const 56l))
     |> print
   in
   Alcotest.(check string)
     "same string"
-    "((aliases_of_canonicals {(x0 {}), (y0 {})})\n\
-    \ (payload_of_canonicals {(x0 (bool.eq x0 42)), (y0 (bool.eq y0 56))}))"
+    "union find:\n\
+    \  ((aliases_of_canonicals {}) (payload_of_canonicals {}))\n\
+     equalities:\n\
+    \  {x0 = 42, y0 = 56}\n"
     pc
 
 let test_print_pc_one_two_sym () : unit =
   let pc =
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (i32_sym "x0") (i32_sym "y0"))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "x0") (i32_sym "y0"))
     |> print
   in
   Alcotest.(check string)
     "same string"
-    "((aliases_of_canonicals {(x0 {y0})})\n\
-    \ (payload_of_canonicals {(x0 (bool.eq x0 y0))}))"
+    "union find:\n\
+    \  ((aliases_of_canonicals {(x0 {y0})})\n\
+    \   (payload_of_canonicals {(x0 (bool.eq x0 y0))}))\n\
+     equalities:\n\
+    \  {}\n"
     pc
 
 let basics_suite =
@@ -69,7 +85,7 @@ let test_slice_on_symbol_single () =
   let slice =
     let x0 = sym_i32 "x0" in
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add
+    |> Symex.Path_condition.add_checked_sat_condition
          (eq (Smtml.Typed.Bitv32.symbol x0) (i32_const 42l))
     |> Symex.Path_condition.slice_on_symbol x0
   in
@@ -81,8 +97,9 @@ let test_slice_on_symbol_irrelevant () =
   let slice =
     let y0 = sym_i32 "y0" in
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (i32_sym "x0") (i32_const 42l))
-    |> Symex.Path_condition.add
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "x0") (i32_const 42l))
+    |> Symex.Path_condition.add_checked_sat_condition
          (eq (Smtml.Typed.Bitv32.symbol y0) (i32_const 56l))
     |> Symex.Path_condition.slice_on_symbol y0
   in
@@ -92,9 +109,11 @@ let test_slice_on_symbol_transitive () =
   let slice =
     let x0 = sym_i32 "x0" in
     let y0 = i32_sym "y0" in
+    let z0 = i32_sym "z0" in
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (Smtml.Typed.Bitv32.symbol x0) y0)
-    |> Symex.Path_condition.add (eq y0 (i32_const 5l))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (Smtml.Typed.Bitv32.symbol x0) y0)
+    |> Symex.Path_condition.add_checked_sat_condition (eq y0 z0)
     |> Symex.Path_condition.slice_on_symbol x0
   in
   Alcotest.(check int)
@@ -109,24 +128,28 @@ let slice_on_symbol_suite =
     ; test_case "transitive" `Quick test_slice_on_symbol_transitive
     ] )
 
-(* `slice_on_condition` *)
+(* `slice_on_new_condition` *)
 
 let test_slice_on_condition_basic () =
   let slice =
     let c1 = eq (i32_sym "x0") (i32_const 10l) in
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add c1
-    |> Symex.Path_condition.slice_on_condition c1
+    |> Symex.Path_condition.add_checked_sat_condition c1
+    |> Symex.Path_condition.slice_on_new_condition c1
   in
   Alcotest.(check int) "condition slice size" 1 (Smtml.Expr.Set.cardinal slice)
 
 let test_slice_on_condition_transitive () =
   let slice =
-    let c1 = eq (i32_sym "x0") (i32_sym "y0") in
+    let x0 = i32_sym "x0" in
+    let y0 = i32_sym "y0" in
+    let z0 = i32_sym "z0" in
+    let c1 = eq x0 y0 in
+    let c2 = eq y0 z0 in
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add c1
-    |> Symex.Path_condition.add (eq (i32_sym "y0") (i32_const 7l))
-    |> Symex.Path_condition.slice_on_condition c1
+    |> Symex.Path_condition.add_checked_sat_condition c1
+    |> Symex.Path_condition.add_checked_sat_condition c2
+    |> Symex.Path_condition.slice_on_new_condition c1
   in
   Alcotest.(check int) "transitive slice size" 2 (Smtml.Expr.Set.cardinal slice)
 
@@ -137,27 +160,31 @@ let slice_on_condition_suite =
     ; test_case "transitive" `Quick test_slice_on_condition_transitive
     ] )
 
-(* `slice` *)
+(* `to_list` *)
 
 let test_slice_empty () =
-  let slices = Symex.Path_condition.empty |> Symex.Path_condition.slice in
+  let slices = Symex.Path_condition.empty |> Symex.Path_condition.to_list in
   Alcotest.(check int) "no slices" 0 (List.length slices)
 
 let test_slice_independent () =
   let slices =
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (i32_sym "x0") (i32_const 1l))
-    |> Symex.Path_condition.add (eq (i32_sym "y0") (i32_const 2l))
-    |> Symex.Path_condition.slice
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "x0") (i32_const 1l))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "y0") (i32_const 2l))
+    |> Symex.Path_condition.to_list
   in
   Alcotest.(check int) "two independent slices" 2 (List.length slices)
 
 let test_slice_connected () =
   let slices =
     Symex.Path_condition.empty
-    |> Symex.Path_condition.add (eq (i32_sym "x0") (i32_sym "y0"))
-    |> Symex.Path_condition.add (eq (i32_sym "y0") (i32_const 3l))
-    |> Symex.Path_condition.slice
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "x0") (i32_sym "y0"))
+    |> Symex.Path_condition.add_checked_sat_condition
+         (eq (i32_sym "y0") (i32_sym "z0"))
+    |> Symex.Path_condition.to_list
   in
   Alcotest.(check int) "one connected slice" 1 (List.length slices)