+++ CHANGES.md
@@ -1,6 +1,8 @@
 ## unreleased
 
+- new API for `Path_condition` which makes enforcing invariants easier
 - expose `Path_condition.pp`
+- substitute equalities in the path condition
 
 ## 0.1 - 2026-02-09